helm: improve quality of Helm chart metadata
Adds a NOTES.txt file, README.md, and provides more details about the
project in Chart.yaml

* https://helm.sh/docs/topics/charts/#the-chart-file-structure
* https://helm.sh/docs/topics/chart_template_guide/notes_files/
* https://helm.sh/docs/topics/charts/#the-chart-yaml-file

The README.md is a generated copy of docs/install-crossplane.md.
This file is inline with what is expected on Helm Hub. We can explore
differentiation between these two uses later.
The frontmatter is removed in the generated copy since chart rendering
tools may not understand it.

In the future we may want to tailor the messaging in the README.md with
more description of what Crossplane is.  The file should act as a
product introduction and helm installation/configuration guide.

Signed-off-by: Marques Johansson <marques@upbound.io>