template stacks: add template stack integration
This adds a first pass at integrating a shared template stack controller
with the stack manager. It also adds a manual integration test and some
instructions to run it.

This includes changes to the data model to support template stacks, in
the form of a new CRD, StackConfiguration. This is expected to converge
with the Stack and StackInstall objects in the future.

The controller consists of a setup phase and a render phase. The render
phase renders a set of configuration into resources and applies them to
a cluster, in response to a CR being created in the cluster. The setup
phase spins up a render phase controller for a single template stack.

The first pass is still experimental and not quite ready for widespread
consumption, so it can only be used by passing `--templates` into the
stack manager invocation.

Signed-off-by: suskin <813571+suskin@users.noreply.github.com>