stacks: expose imagePullPolicy, imagePullSecrets, and SA Annotations in StackInstall
The ServiceAccount created to run the Stack controller deployment will include
the specified annotations.

The Deployment created for the Stack controller will include the specified
imagePullPolicy and imagePullSecrets.

The initContainer of the Job used by the StackManager, which copies the stack
package to a shared volume, will use the imagePullPolicy and imagePullSecrets
specified in the StackInstall to access the stack image.

The stack unpack processing container of this Job will, however, rely on the
same imagePullPolicy used on the StackManager.

Signed-off-by: Marques Johansson <marques@upbound.io>