stacks: add stack-manager option to force image pull policy
In multi-tenant environments, images fetched for one Stack unpack or Stack
controller become available in the node level docker registry. This could
allow other tenants to access those images without the appropriate
imagePullSecrets.

This change introduces a stack-manager parameter: `--force-image-pull-policy`,
which when set to a non-empty string, will override the imagePullPolicy set in
any StackInstall spec (affecting the install job and upack) and will force an
imagePullPolicy for each pod in a Stack deployment.

Signed-off-by: Marques Johansson <marques@upbound.io>