Vendor OAM types and generate CRDs on make generate
Now that the structs for OAM types has been moved
to oam-kubernetes-remote, we must vendor them and
generate the CRDs from the vendored dir. The vendor
directory is not committed, but we ensure that the
generated CRDs match by including vendoring in the
make generate command, which will be cheked by
make check-diff on builds.

Signed-off-by: hasheddan <georgedanielmangum@gmail.com>