Allow composed resources to specify inline connection details
This commit allows a composed resource to specify the name and value of a
connection detail that should be aggregated up to the composite resource.
One example of this pattern being useful is for our getting started guide,
which creates a PostgreSQL database using Alibaba, AWS, Azure, and GCP. Of
these clouds only AWS and Alibaba include a port in their connection secret.

As well as helping compositions satsify a definition in a portable fashion,
this allows composition authors to expose fixed, non-sensitive connection
details (e.g. ports, regions, protocols, database versions) about their composed
resources via a secret that can be loaded by a pod as env vars or config files.

Signed-off-by: Nic Cope <negz@rk0n.org>