Rename InfrastructureDefinition to CompositeResourceDefinition
Originally we had planned to have both application and infrastructure focused
composite resources. We've since changed our approach to applications, and thus
"composite resource" and "composite infrastructure resource" have become
synonymous. We tend to say "composite resource" by convention, and thus
"composite resource definition" fits as the definition of a composite resource.
We're using 'XRD' as the abbreviation to avoid confusion with CRDs.

This commit also renames InfrastructurePublication to
CompositeResourcePublication. We plan to roll this concept into the XRD rather
than a distinct resource in a future commit.

Signed-off-by: Nic Cope <negz@rk0n.org>