Remove the original, superceded, package manager
We've taken our learnings from the original Crossplane package manager, and used
them to build a new one. The original package manager grew from the 'stack
manager'. It most recently operated on types in the packages.crossplane.io API
group - e.g. ClusterPackageInstall. This package manager was designed to solve
problems that Crossplane now deems to be outside its scope - i.e. installing
user-facing application logic, and arbitrary Kubernetes controllers.

The new package manager operates on types in the pkg.crossplane.io API group. It
is designed specifically to install providers and configuration for Crossplane,
not arbitrary controllers or applications. The new package manager is designed
to support updating packages (e.g. to install a new version of a provider) and
to automatically resolve package dependencies. (The latter is coming soon.)

The removal of resource claims in v0.13 mean the majority of v0.12 (and earlier)
era packages will not function with v0.13 due to the lack of claim and class
CRDs:

* Provider packages will panic without claim CRDs installed.
* Many app packages (e.g. Wordpress) will not function without claim CRDs

Upgrading to v0.13 without also migrating to v0.13 era providers (and thus to
the v0.13 package manager) may be possible with manual intervention, but is
likely to be equally as or more complicated than upgrading to v0.13 and to the
new package manager.

Signed-off-by: Nic Cope <negz@rk0n.org>