Replace image identfier in helm provider names
Replaces all characters following a : or @ in a provider image that is
passed via provider.packages when setting the provider name. This more
closely aligns with the behavior of crank and does not include version
information in the provider name as the provider may be updated to a
different version in the future.

Signed-off-by: hasheddan <georgedanielmangum@gmail.com>