Make dependency status fields optional in revisions
In rare edge cases, having required status fields can cause problems
when trying to clean up old objects by manually removing finalizers.
kubectl is not able to edit the status subresource, meaning that if
these fields are not set then a user is stuck trying to remove the
finalizer because the patch is technically invalid. This would only be
the case if a user was updating from a pre-v1.0.0 Crossplane version and
had manually upgraded the revision CRDs to v1.

Signed-off-by: hasheddan <georgedanielmangum@gmail.com>