Allow Composition authors to reorder resource templates
This commit allows Compositions to name their templates. When a Composition's
templates are named Crossplane is able to account for more kinds of update to
the template array. The array can be reordered, and elements can be added and
deleted regardless of their index within the array. When a named template is
deleted any composed resource that was created using that template is also
deleted.

Either all or no templates must be named. When no templates are named Crossplane
behaves as it did before this commit was introduced; i.e. it allows templates to
be appended to the templates array, but does not support reordering or deletion.

Signed-off-by: Nic Cope <negz@rk0n.org>