Rename 'templateName' to 'name' in the Composition resources array
I was initially reluctant to use 'name' because I worried that folks would
interpret it as specifying the metadata.name of the composed resource, rather
than just uniquely identifying the entry in the resources array. In the end I
decided that I prefer 'name' to alternatives like 'id', 'templateName', or
'baseName'. API conventions prefer lists of 'named' subobjects over maps, and
imply that the name field should be named 'name'. This is pretty common in
Kubernetes APIs, so hopefully it won't be too confusing for folks.

Signed-off-by: Nic Cope <negz@rk0n.org>