Add base for controller config support (#10)
Towards: https://github.com/giantswarm/giantswarm/issues/24737

What I primarily try to solve here is to be able to enable debug logging
for example for the AWS operator but controller configs are generally
vey nice to have better control over the deployer provider operators.

I have one security concern with the implementation, which may not be
valid cos the added counter measure + the fact that we manage it via
Flux in the MCs. Please check my comment below.